Adding multi-select input and improving input arguments
Creating an option to include a menu input in the form. Also improving input arguments when creating the form, requiring the user to include the name of the input, the input type, and another other additional arguments rather than just a name. Currently need to fix handleChange() function to avoid redundant code and implement other possible input options for other types of forms.